requirements.txt
alembic==1.10.4
appdirs==1.4.4
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
fake-useragent==1.1.3
filelock==3.12.0
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
parse==1.19.0
parsel==1.8.1
Protego==0.2.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==8.2.2
pyOpenSSL==23.1.1
pyppeteer==1.0.2
pyquery==2.0.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
requests-html==0.10.0
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.10
tldextract==3.4.0
tqdm==4.65.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
Werkzeug==2.2.3
zipp==3.15.0
zope.interface==6.0
